# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management API
  description: |
    REST API для управления банковскими картами,
    пользователями, запросами на блокировку
    и переводами между собственными картами.
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/admin/users/{userId}:
    get:
      tags:
      - Пользователи
      summary: Получить пользователя
      description: Возвращает пользователя по его идентификатору
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        description: Идентификатор пользователя
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Пользователь найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав для выполнения операции
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Пользователи
      summary: Обновить пользователя
      description: Изменяет данные существующего пользователя
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: Идентификатор пользователя
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: Данные пользователя успешно обновлены
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Переданы некорректные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Новый логин уже занят
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Пользователи
      summary: Удалить пользователя
      description: Удаляет пользователя по его идентификатору
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: Идентификатор пользователя
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: Пользователь успешно удалён
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Пользователь не может быть удалён из-за связанных данных
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
  /api/admin/users:
    get:
      tags:
      - Пользователи
      summary: Получить список пользователей
      description: Возвращает страницу пользователей с поддержкой пагинации и сортировки
      operationId: getAllUsers
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Список пользователей успешно получен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав для выполнения операции
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Пользователи
      summary: Создать пользователя
      description: Создаёт нового пользователя с ролью USER или ADMIN
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Переданы некорректные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав для выполнения операции
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Пользователь с таким логином уже существует
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
    UpdateUserRequest:
      type: object
      description: Данные для обновления пользователя
      properties:
        firstName:
          type: string
          description: Новое имя пользователя
          example: Иван
          maxLength: 50
          minLength: 0
        middleName:
          type: string
          description: Новое отчество пользователя
          example: Иванович
          maxLength: 50
          minLength: 0
        lastName:
          type: string
          description: Новая фамилия пользователя
          example: Иванов
          maxLength: 50
          minLength: 0
        role:
          type: string
          description: Новая роль пользователя
          enum:
          - ADMIN
          - USER
          example: USER
      required:
      - firstName
      - lastName
      - role
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки REST API
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2026-07-28T19:34:52.426
        status:
          type: integer
          format: int32
          description: HTTP-статус ошибки
          example: 404
        error:
          type: string
          description: Наименование HTTP-ошибки
          example: Not Found
        message:
          type: string
          description: Сообщение для пользователя
          example: Карта с идентификатором 15 не найдена
        path:
          type: string
          description: Путь HTTP-запроса
          example: /api/cards/15
    UserResponse:
      type: object
      description: Информация о пользователе
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор пользователя
          example: 1
        firstName:
          type: string
          description: Имя пользователя
          example: Иван
        middleName:
          type: string
          description: Отчество пользователя
          example: Иванович
        lastName:
          type: string
          description: Фамилия пользователя
          example: Иванов
        login:
          type: string
          description: Логин пользователя
          example: ivanov
        role:
          type: string
          description: Роль пользователя
          enum:
          - ADMIN
          - USER
          example: USER
...
    CreateUserRequest:
      type: object
      description: Данные для создания пользователя
      properties:
        firstName:
          type: string
          description: Имя пользователя
          example: Иван
          maxLength: 50
          minLength: 0
        middleName:
          type: string
          description: Отчество пользователя
          example: Иванович
          maxLength: 50
          minLength: 0
        lastName:
          type: string
          description: Фамилия пользователя
          example: Иванов
          maxLength: 50
          minLength: 0
        login:
          type: string
          description: Логин пользователя
          example: ivanov
          maxLength: 50
          minLength: 3
        password:
          type: string
          description: Пароль пользователя
          example: Password123
          maxLength: 100
          minLength: 8
        role:
          type: string
          description: Роль пользователя
          enum:
          - ADMIN
          - USER
          example: USER
      required:
      - firstName
      - lastName
      - login
      - password
      - role
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT access-токен для авторизации
      scheme: bearer
      bearerFormat: JWT